// src/app/pages/cabinet/enums/institute.enum.ts
export enum InstituteEnum {
  Vyc = 'Vyc', // Военный учебный центр
  Ienim = 'Ienim', // Институт естественных наук и математики
  Inmit = 'Inmit', // Институт новых материалов и технологий
  Rtf = 'Rtf', // Институт радиоэлектроники и информационных технологий–РтФ
  Isa = 'Isa', // Институт строительства и архитектуры
  Itoo = 'Itoo', // Институт технологий открытого образования
  Ifk = 'Ifk', // Институт физической культуры, спорта и молодежной политики
  Ifo = 'Ifo', // Институт фундаментального образования
  Iey = 'Iey', // Институт экономики и управления
  Ygi = 'Ygi', // Уральский гуманитарный институт
  Yei = 'Yei', // Уральский энергетический институт
  Fti = 'Fti', // Физико-технологический институт
  Hti = 'Hti' // Химико-технологический институт
}